renamed 'ingestor' to 'eif_ingestor'. removed the (unused) check_doi feature
